+++ src/org/opensha2/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Entry points for conducting probabilisitic seismic hazard analysis (PSHA) calculations.
  */
-package org.opensha2.programs;
+package org.opensha2;
